input avatar refactor
