+++ src/components/ModalCrop/ModalCrop.module.scss
@@ -1,0 +1,16 @@
+.ModalCrop {
+    align-items: center;
+    background-color: rgba(0, 0, 0, 0.85);
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    left: 0;
+    min-height: 100vh;
+    position: absolute;
+    top: 0;
+    width: 100vw;
+
+    img {
+        max-height: 90vh;
+    }
+}
