+++ src/views/SignIn/styles.scss
@@ -17,37 +17,6 @@
             flex-wrap: wrap;
             width: 60%;
         }
-        &__image-wrapper {
-            display: flex;
-            flex-wrap: wrap;
-            justify-content: center;
-            margin: auto;
-            width: 30%;
-
-            &__image {
-                border-radius: 10px;
-                height: auto;
-                width: 100%;
-            }
-
-            &__input {
-                &--hide {
-                    display: none;
-                }
-            }
-
-            &__button {
-                border-radius: 5px;
-                font-size: 16px;
-                padding: 2px 0;
-                white-space: normal;
-                width: 100%;
-
-                &:hover {
-                    cursor: pointer;
-                }
-            }
-        }
     }
 }
